Fix SMSC 911x Ethernet EEPROM to work properly with SEAD3.
The SEAD3 does not have an EEPROM to store the MAC address, so we should
set the flag for the SMSC 911x Ethernet driver accordingly. We also do
not issue a soft reset in this case when the device is opened.

Signed-off-by: Steven J. Hill <sjhill@mips.com>
(cherry picked from commit c4b3eabd99a816974ebb816592b16e8f278f5d3f)