Fix micro-assembled jump immediate for microMIPS.
The micro-assembled jump immediate for microMIPS is incorrect. This
produced a warning message of 'Micro-assembler field overflow' that
is now a panic and prevents the kernel from booting. The check is
disabled for microMIPS which can jump to aligned addresses.

Signed-off-by: Steven J. Hill <sjhill@mips.com>
(cherry picked from commit 15ce386bf90ca3469b2e46e6299d371069b3747b)