Partial revert of WARN() is uasm for microMIPS kernel.
The original patch in commit 8d662c8d34a05e8e47deaa9e22fe770dc557c2d3
changes the printks to WARNs to instead produce a stack trace. For a
pure microMIPS kernel, the WARN is disabled for when a jump immediate
instruction is built. This is a temporary hack until the TLB exceptions
get converted to be pure microMIPS instructions.

Signed-off-by: Steven J. Hill <sjhill@mips.com>