MIPS: Malta: check memory map type - legacy or new
Malta board has two memory map variants - legacy and new and
there are a couple of possible and impossible variants between
kernel configuration and system controller configuration.
That stuff really troubles!

So - read a secret sysctrl register, check and do decision:
continue or not or continue with memory restiction to 256MB.

Change-Id: I3a7ab4fe9a57b59e882140d757f2c394e7a4a949
Author: Leonid Yegoshin <yegoshin@mips.com>
Signed-off-by: Leonid Yegoshin <yegoshin@mips.com>(imported from commit dcde2b9df2b626bcd55d3cbd683a0c66874abeed)