MIPS: (opt) Fix of reading I-pages from non-DMA FS devices for ID cache separation
This optional fix provides a D-cache flush for instruction code pages on
page faults. In case of non-DMA block device a driver doesn't know that it
reads I-page and doesn't flush D-cache generally on systems without
cache aliasing. And that takes toll during page fault of instruction pages.

It is not a perfect fix, it should be considered as a temporary fix.
The permanent fix would track pages in page cache and flushes D-cache
during receiption of page from driver only but not at each page fault.
It is not done yet.

Change-Id: I43f5943d6ce0509729179615f6b81e77803a34ac
Author: Leonid Yegoshin <yegoshin@mips.com>
Signed-off-by: Leonid Yegoshin <yegoshin@mips.com>(imported from commit 6ebd22eb7a3d9873582ebe990a77094f971652ee)(imported from commit 0caf3b4a1eebb64572e81e4df6fdb3abf12c7017)