MIPS: FPU2 IEEE754-2008 SNaN support
(FPU2 supports only IEEE754-2008 Nan/SNaN formats but not MIPS)

Because IEEE754-2008 SNaN single precision is not compatible with
IEEE double precision SNaN, we can't use the single DP value to cover both, so
there is no win-win situation for -mfp32 applications.

So, FPRs are initialized for MIPS SNaN for FPU1 or -mfp32 applications.
In case of FPU2 and -mfp64 application (32 regs by 64 bits) IEEE SNaN
is used for FPRs initialization. I know, it is not ideal for GCCv4+ -mfp32,
but it still has IEEE2008 QNaN. At least FPRs initialization is a tool
to trigger bugs in GCC and kernel but not in applications.

Also, kernel FPU emulator is fixed to support both NaN variants.
It matches FPU NaN capability (looking into FCSR31.NAN2008 bit).
(But it is no uMIPS32 -mfp64 for FPU2 support yet).

Kernel boot parameter 'nan2008=0/1' can be used to enforce selection
in case of non-FPU cores (in kernel FPU emulator).

Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>
Signed-off-by: Steven J. Hill <Steven.Hill@imgtec.com>