MIPS: bugfix of coherentio variable default setup
Patch

commit 92980ac184430a1a097c87b30fd8193c02e57f61
Author: Steven J. Hill <Steven.Hill@imgtec.com>
Date:   Wed Mar 27 00:12:23 2013 -0500

    MIPS: Add option to disable software I/O coherency.

    Some MIPS controllers have hardware I/O coherency. This patch
    detects those and turns off software coherency. A new kernel
    command line option also allows the user to manually turn
    software coherency on or off.

    Signed-off-by: Steven J. Hill <Steven.Hill@imgtec.com>

in fact enforces L2 cache flushes even on systems with IOCU.
The default value of coherentio is 0 and is not changed even with IOCU.
It is a serious performance problem because it destroys all IOCU performance
advantages.

It is fixed by setting default value as (-1) and setup a final value
during platform coherency setup.

Author: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>
Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>