MIPS: bugfix of force-bev-location parsing
Parsing of force-bev-location was incorrect.
The code was originally written to support YAMON environment variables
of the form "force-bev-location=<address>", but Linux automatically
strips off the "force-bev-location=" when used with early_param().
Additionally it can be called from BEV_overlay_segment() straight from
strstr() which may return NULL so it needs to handle that case too.

Author: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>
Signed-off-by: Leonid Yegoshin <Leonid.Yegoshin@imgtec.com>