mips: default NR_CPUS=8 for malta SMP defconfigs
The previous NR_CPUS=2 default is not an optimal default for current
Malta setups where it is common to have more than 2 CPUs available. It
makes sense to increase this to a number which covers all common setups
currently in use, such that all of those cores are usable. 8 seems to
fit that description.

Additionally if a current kernel runs on a bitstream with more CPUs (be
they cores or VPEs) then the kernel will repeatedly print the following
error:

 unexpected IRQ # 52

...and the kernel will appear to lock up as it deals with the repeated
interrupt. Note that this patch is in no way a fix for that issue,
working around it for the moment is just a nice side affect.

If the user has less than 8 CPUs & they wish to have a more optimal
kernel they can simply reduce this in their config. It makes sense for
the default to work on as many systems as possible.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>