MIPS: move GIC IPI functions out of smp-cmp.c
The GIC IPI functions aren't necessarily specific to the "CMP
framework" SMP implementation, and will be used elsewhere in a
subsequent commit. This patch moves them to a separate file which is
compiled when a new CONFIG_MIPS_GIC_IPI Kconfig symbol is selected, and
selects that symbol for the current MIPS_CMP user.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>