MIPS: Add syscall_set_return_value system call
The syscall_set_return_value sets the return value for the
executed system call. According to the MIPS syscall calling
conventions, the return value is 0 if the syscall was executed
successfully or non-0 if the syscall encountered a problem. A negative
errno must also be returned to userland in case of a syscall failure.

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>