MIPS: asm: uaccess: Fix EVA #ifdef for __strncpy_from_user
__strncpy_from_user is used to copy a NULL terminated string from
userland with less checking. However, the code was wrong because in
non-EVA mode the __strncpy_from_user_nocheck_asm was used which is
only defined for EVA in arch/mips/lib/strncpy_user.S. If the code
ever tried to jump to this function (declared as 'extern') it would
end up in an undefined location. We fix this problem by using this
function in EVA mode and use the __strncpy_from_kernel_nocheck_asm
in non-EVA one.

Signed-off-by: Markos Chandras <markos.chandras@imgtec.com>