cpuidle: coupled: disable IRQs after leaving safe state
If a CPU enters a safe state whilst it waits for coupled CPUs to become
ready to enter a deeper coupled idle state then the cpuidle_enter_state
call used to enter the safe state will enable IRQs. Since commit
736899ab702d "cpuidle: coupled: fix race condition between pokes and
safe state" whether or not IRQs remain enabled following the call to
cpuidle_coupled_clear_pokes is dependant upon whether or not a poke was
pending. This leads to the state of IRQs not being consistent when
entering either the safe state again or entering a coupled idle state.

Disable IRQs such that they are consistently disabled when entering idle
states.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>