MIPS: Malta: Pseudo suspend to RAM
Implement Pseudo suspend to RAM for Malta as a proof of concept. This
isn't proper suspend to RAM as the resume address isn't written to any
scratch registers. Instead it is printed to the console prior to
suspend, which actually restarts the machine rather than entering a
low power state. The user must then re-enter the kernel manually from
the bootloader prompt.

This also includes a hack to disable freeing of YAMON memory, just to
make sure it doesn't get reused for something important before YAMON
clobbers it during resume.

Signed-off-by: James Hogan <james.hogan@imgtec.com>