MIPS: irq-gic: Add Cluster PM callbacks to preserve GIC state
Implement CPU cluster power management callbacks for the GIC, primarily
for suspend to RAM. Mostly hardware state is directly saved/restored,
except for the mappings to pin and vpe, where the registers are
described as write-only, so a shadow copy is kept in memory to allow
them to be restored on cluster power up.

This includes preservation of the configuration done by the cevt-gic
clock event driver to route the count/compare interrupt.

Signed-off-by: James Hogan <james.hogan@imgtec.com>