MIPS: PM: Implement PM helper macros
Implement assembler helper macros in asm/pm.h for platform code to use
for suspend to RAM. Macros are provided for saving and restoring the
main CPU context used by C code and doing important configuration which
must be done very early during resume. Notably EVA needs segmentation
control registers to be restored before the stack or dynamically
allocated memory is accessed, so that state is saved in global data.

Signed-off-by: James Hogan <james.hogan@imgtec.com>