MIPS: cpuidle-cps: rename nc_ready_count to ready_count
This variable indicates the number of VPEs within a core which are ready
to enter a non-coherent state. However it is not a non-coherent mapping
of that variable, nor will it be specific to the non-coherent wait idle
state. Rename it to avoid confusion and allow the name nc_ready_count to
be used in future to refer to a non-coherent mapping of ready_count.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>