MIPS: cpuidle-cps: support clock gating cores
This patch implements a new idle state in which the clock of an idle
core is gated. This removes all dynamic power usage from the core and
leads the core to cease executing instructions or servicing interrupts.
Static power remains preserving the register state of the core.

Because the core becomes incapable of servicing interrupts its clock
must be restarted explicitly by another core. Code is added to
gic_send_ipi_single which will issue a CPC power-up command to achieve
this when appropriate.

When the core clock is gated its CP0 counter naturally stops counting,
and because the core is unable to service interrupts its local GIC
counter is ineffective too. Thus clock gated cores must make use of the
broadcast timer, and the system must therefore include a clock event
driver suitable for use as the broadcast timer.

There must always be at least one core which is powered up & clocked
(ie. not in a clock-off idle state or deeper). To satisfy that
requirement core 0 is currently never allowed to enter an idle state
deeper than non-coherent wait.

Interrupt affinity is not modified on entry to the clock-off idle state,
so any interrupts whose affinity is mapped to the clock gated core will
not be serviced until that core is re-clocked due to an IPI. For Malta
this does not currently matter since all hardware interrupts are
serviced by cpu0 (which being in core 0 never enters the clock-off idle
state).

Signed-off-by: Paul Burton <paul.burton@imgtec.com>