MIPS: pm-cps: don't bother clearing ready_count on nc-wait exit
Since clock gating support was implemented the ready_count variable is
cleared on entry to non-coherent idle states, since the clock gating
state has no need to count VPEs as they exit the state. This leads the
the clearing of the top bit of ready_count by the final VPE to leave
the non-coherent wait state being redundant. Remove it.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>