MIPS: smp-cps: do MT core init in asm, simplify it
This patch moves initialisation of secondary VPEs within a core to an
assembly function called from the boot vector. This will become
important once hotplug is implemented if a secondary VPE is the first to
be brought online within a core. In that case the first VPE of the core
will start & needs to be able to initialise the secondary VPEs, run the
appropriate one & halt itself. This all needs to be done potentially
without a valid stack pointer or gp value, which is made much easier by
doing it in assembly.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>