MIPS: cpuidle-cps: support core power down
This patch introduces support for a new idle state which power gates
idle cores. This removes all static & dynamic power from the core and it
thus consumes no power whilst in this idle state, however all register
state is lost and must therefore be saved & restored around the power
gated state, in addition to reinitialising certain core-level hardware
such as the L1 caches. The CPS SMP implementation must be in use for
the power gated PM state, and thus the power gated idle state, to be
available.

Signed-off-by: Paul Burton <paul.burton@imgtec.com>