dynamic_debug: refactor dynamic_pr_debug and friends
For the upcoming 'define the _ddebug descriptor in assembly', we need all
the descriptors in a translation unit to have distinct names (because asm
does not understand C scope).  The easiest way to achieve that is as usual
with an extra level of macros, passing the identifier to use to the
innermost macro, generating it via __UNIQUE_ID or something.

However, instead of repeating that exercise for dynamic_pr_debug,
dynamic_dev_dbg, dynamic_netdev_dbg and dynamic_hex_dump separately, we
can use the similarity between their bodies to implement them via a common
macro, _dynamic_func_call - though the hex_dump case requires a slight
variant, since print_hex_dump does not take the _ddebug descriptor.  We'll
also get to use that variant elsewhere (btrfs).

Link: http://lkml.kernel.org/r/20190212214150.4807-11-linux@rasmusvillemoes.dk
Signed-off-by: Rasmus Villemoes <linux@rasmusvillemoes.dk>
Acked-by: Jason Baron <jbaron@akamai.com>
Cc: David Sterba <dsterba@suse.com>
Cc: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Cc: Ingo Molnar <mingo@kernel.org>
Cc: Petr Mladek <pmladek@suse.com>
Cc: "Rafael J . Wysocki" <rafael.j.wysocki@intel.com>
Cc: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>